+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - New config: `svelte` via [`eslint-plugin-svelte`](https://www.npmjs.com/package/eslint-plugin-svelte), enabled automatically if `svelte` package is installed.
 - New config: `es` via [`eslint-plugin-es-x`](https://www.npmjs.com/package/eslint-plugin-es-x), **<u>disabled</u>** by default.
+- Added an option to `import` config to specify `eslint-plugin-import-x` plugin settings.
 
 ## 0.9.0